+++ library/src/main/java/com/leinardi/android/speeddial/FabWithLabelView.java
@@ -16,6 +16,7 @@
 
 package com.leinardi.android.speeddial;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.content.res.TypedArray;
@@ -81,6 +82,7 @@ public class FabWithLabelView extends LinearLayout {
         init(context, attrs);
     }
 
+    @SuppressLint("RestrictedApi")
     @Override
     public void setVisibility(int visibility) {
         super.setVisibility(visibility);
@@ -117,28 +119,38 @@ public class FabWithLabelView extends LinearLayout {
     }
 
     /**
-     * Return This method returns fab labels background card.
+     * Returns FAB labels background card.
      */
     public CardView getLabelBackground() {
         return mLabelCardView;
     }
 
     /**
-     * Return This method returns fab in action menu.
+     * Returns the {@link FloatingActionButton}.
      */
     public FloatingActionButton getFab() {
         return mFab;
     }
 
-    @Nullable
     public SpeedDialActionItem getSpeedDialActionItem() {
+        if (mSpeedDialActionItem == null) {
+            throw new IllegalStateException("SpeedDialActionItem not set yet!");
+        }
         return mSpeedDialActionItem;
+    }
+
+    /**
+     * Returns an instance of the {@link SpeedDialActionItem.Builder} initialized with the current instance of the
+     * {@link SpeedDialActionItem} to make it easier to modify the current Action Item settings.
+     */
+    public SpeedDialActionItem.Builder getSpeedDialActionItemBuilder() {
+        return new SpeedDialActionItem.Builder(getSpeedDialActionItem());
     }
 
     public void setSpeedDialActionItem(SpeedDialActionItem actionItem) {
         mSpeedDialActionItem = actionItem;
         setId(actionItem.getId());
-        setLabel(actionItem.getLabel());
+        setLabel(actionItem.getLabel(getContext()));
         SpeedDialActionItem speedDialActionItem = getSpeedDialActionItem();
         setLabelClickable(speedDialActionItem != null && speedDialActionItem.isLabelClickable());