+++ library/src/main/java/com/leinardi/android/speeddial/SpeedDialActionItem.java
@@ -24,9 +24,11 @@ import android.support.annotation.ColorInt;
 import android.support.annotation.DrawableRes;
 import android.support.annotation.IdRes;
 import android.support.annotation.Nullable;
+import android.support.annotation.StringRes;
 import android.support.annotation.StyleRes;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.content.res.AppCompatResources;
+import android.view.ContextThemeWrapper;
 
 import static android.support.design.widget.FloatingActionButton.SIZE_AUTO;
 
@@ -37,6 +39,8 @@ public class SpeedDialActionItem implements Parcelable {
     private final int mId;
     @Nullable
     private final String mLabel;
+    @StringRes
+    private final int mLabelRes;
     @DrawableRes
     private final int mFabImageResource;
     @Nullable
@@ -58,6 +62,7 @@ public class SpeedDialActionItem implements Parcelable {
     private SpeedDialActionItem(Builder builder) {
         mId = builder.mId;
         mLabel = builder.mLabel;
+        mLabelRes = builder.mLabelRes;
         mFabImageTintColor = builder.mFabImageTintColor;
         mFabImageResource = builder.mFabImageResource;
         mFabImageDrawable = builder.mFabImageDrawable;
@@ -74,8 +79,14 @@ public class SpeedDialActionItem implements Parcelable {
     }
 
     @Nullable
-    public String getLabel() {
-        return mLabel;
+    public String getLabel(Context context) {
+        if (mLabel != null) {
+            return mLabel;
+        } else if (mLabelRes != RESOURCE_NOT_SET) {
+            return context.getString(mLabelRes);
+        } else {
+            return null;
+        }
     }
 
     /**
@@ -123,6 +134,18 @@ public class SpeedDialActionItem implements Parcelable {
         return mTheme;
     }
 
+    public FabWithLabelView createFabWithLabelView(Context context) {
+        FabWithLabelView fabWithLabelView;
+        int theme = getTheme();
+        if (theme == RESOURCE_NOT_SET) {
+            fabWithLabelView = new FabWithLabelView(context);
+        } else {
+            fabWithLabelView = new FabWithLabelView(new ContextThemeWrapper(context, theme), null, theme);
+        }
+        fabWithLabelView.setSpeedDialActionItem(this);
+        return fabWithLabelView;
+    }
+
     // Disabled due to https://issuetracker.google.com/issues/77303906
     @FloatingActionButton.Size
     /* public */ int getFabSize() {
@@ -140,6 +163,8 @@ public class SpeedDialActionItem implements Parcelable {
         private int mFabImageTintColor = RESOURCE_NOT_SET;
         @Nullable
         private String mLabel;
+        @StringRes
+        private int mLabelRes = RESOURCE_NOT_SET;
         @ColorInt
         private int mFabBackgroundColor = RESOURCE_NOT_SET;
         @ColorInt
@@ -153,7 +178,7 @@ public class SpeedDialActionItem implements Parcelable {
         private int mTheme = RESOURCE_NOT_SET;
 
         /**
-         * Creates a builder for a speed dial action item that uses the a {@link DrawableRes} as icon.
+         * Creates a builder for a speed dial action item that uses a {@link DrawableRes} as icon.
          *
          * @param id               the identifier for this action item. The identifier must be unique to the instance
          *                         of {@link SpeedDialView}. The identifier should be a positive number.
@@ -166,7 +191,7 @@ public class SpeedDialActionItem implements Parcelable {
         }
 
         /**
-         * Creates a builder for a speed dial action item that uses the a {@link Drawable} as icon.
+         * Creates a builder for a speed dial action item that uses a {@link Drawable} as icon.
          * <p class="note">{@link Drawable} are not parcelables so is not possible to restore them when the view is
          * recreated for example after an orientation change. If possible always use the {@link #Builder(int, int)}</p>
          *
@@ -180,8 +205,34 @@ public class SpeedDialActionItem implements Parcelable {
             mFabImageResource = RESOURCE_NOT_SET;
         }
 
+        /**
+         * Creates a builder for a speed dial action item that uses a {@link SpeedDialActionItem} instance to
+         * initialize the default values.
+         *
+         * @param speedDialActionItem it will be used for the default values of the builder.
+         */
+        public Builder(SpeedDialActionItem speedDialActionItem) {
+            mId = speedDialActionItem.mId;
+            mLabel = speedDialActionItem.mLabel;
+            mLabelRes = speedDialActionItem.mLabelRes;
+            mFabImageResource = speedDialActionItem.mFabImageResource;
+            mFabImageDrawable = speedDialActionItem.mFabImageDrawable;
+            mFabImageTintColor = speedDialActionItem.mFabImageTintColor;
+            mFabBackgroundColor = speedDialActionItem.mFabBackgroundColor;
+            mLabelColor = speedDialActionItem.mLabelColor;
+            mLabelBackgroundColor = speedDialActionItem.mLabelBackgroundColor;
+            mLabelClickable = speedDialActionItem.mLabelClickable;
+            mFabSize = speedDialActionItem.mFabSize;
+            mTheme = speedDialActionItem.mTheme;
+        }
+
         public Builder setLabel(@Nullable String label) {
             mLabel = label;
+            return this;
+        }
+
+        public Builder setLabel(@StringRes int labelRes) {
+            mLabelRes = labelRes;
             return this;
         }
 
@@ -236,6 +287,7 @@ public class SpeedDialActionItem implements Parcelable {
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeInt(this.mId);
         dest.writeString(this.mLabel);
+        dest.writeInt(this.mLabelRes);
         dest.writeInt(this.mFabImageResource);
         dest.writeInt(this.mFabImageTintColor);
         dest.writeInt(this.mFabBackgroundColor);
@@ -249,6 +301,7 @@ public class SpeedDialActionItem implements Parcelable {
     protected SpeedDialActionItem(Parcel in) {
         this.mId = in.readInt();
         this.mLabel = in.readString();
+        this.mLabelRes = in.readInt();
         this.mFabImageResource = in.readInt();
         this.mFabImageDrawable = null;
         this.mFabImageTintColor = in.readInt();