+++ library/src/main/java/com/leinardi/android/speeddial/SpeedDialView.java
@@ -45,7 +45,6 @@ import android.support.v7.widget.PopupMenu;
 import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.util.Log;
-import android.view.ContextThemeWrapper;
 import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -56,6 +55,9 @@ import android.view.animation.AnimationUtils;
 import android.widget.LinearLayout;
 
 import java.lang.annotation.Retention;
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
@@ -179,7 +181,47 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
 
     public void show(@Nullable final OnVisibilityChangedListener listener) {
         setVisibility(VISIBLE);
-        mMainFab.show(listener);
+        showFabWithWorkaround(mMainFab, listener);
+    }
+
+    /*
+     * WORKAROUND: Remove if Google will finally fix this: https://issuetracker.google.com/issues/111316656
+     */
+    private void showFabWithWorkaround(FloatingActionButton fab, @Nullable final OnVisibilityChangedListener listener) {
+        fab.show(new OnVisibilityChangedListener() {
+            @SuppressWarnings("unchecked")
+            @Override
+            public void onShown(FloatingActionButton fab) {
+                try {
+                    Field declaredField = fab.getClass().getDeclaredField("impl");
+                    declaredField.setAccessible(true);
+                    Object impl = declaredField.get(fab);
+                    Class implClass = Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP
+                            ? impl.getClass().getSuperclass() : impl.getClass();
+                    Method scale = implClass.getDeclaredMethod("setImageMatrixScale", Float.TYPE);
+                    scale.setAccessible(true);
+                    scale.invoke(impl, 1.0F);
+                } catch (NoSuchMethodException e) {
+                    Log.e(TAG, "Method setImageMatrixScale not found", e);
+                } catch (IllegalAccessException e) {
+                    Log.e(TAG, "IllegalAccessException", e);
+                } catch (InvocationTargetException e) {
+                    Log.e(TAG, "InvocationTargetException", e);
+                } catch (NoSuchFieldException e) {
+                    Log.e(TAG, "Field impl not found", e);
+                }
+                if (listener != null) {
+                    listener.onShown(fab);
+                }
+            }
+
+            @Override
+            public void onHidden(FloatingActionButton fab) {
+                if (listener != null) {
+                    listener.onHidden(fab);
+                }
+            }
+        });
     }
 
     public void hide() {
@@ -265,20 +307,25 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
      * collection's Iterator.
      *
      * @param actionItemCollection collection containing {@link SpeedDialActionItem} to be added to this list
+     * @return a collection containing the instances of {@link FabWithLabelView} added.
      */
-    public void addAllActionItems(Collection<SpeedDialActionItem> actionItemCollection) {
+    public Collection<FabWithLabelView> addAllActionItems(Collection<SpeedDialActionItem> actionItemCollection) {
+        ArrayList<FabWithLabelView> fabWithLabelViews = new ArrayList<>();
         for (SpeedDialActionItem speedDialActionItem : actionItemCollection) {
-            addActionItem(speedDialActionItem);
+            fabWithLabelViews.add(addActionItem(speedDialActionItem));
         }
+        return fabWithLabelViews;
     }
 
     /**
      * Appends the specified {@link SpeedDialActionItem} to the end of this list.
      *
      * @param speedDialActionItem {@link SpeedDialActionItem} to be appended to this list
+     * @return the instance of the {@link FabWithLabelView} if the add was successful, null otherwise.
      */
-    public void addActionItem(SpeedDialActionItem speedDialActionItem) {
-        addActionItem(speedDialActionItem, mFabWithLabelViews.size());
+    @Nullable
+    public FabWithLabelView addActionItem(SpeedDialActionItem speedDialActionItem) {
+        return addActionItem(speedDialActionItem, mFabWithLabelViews.size());
     }
 
     /**
@@ -287,9 +334,11 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
      *
      * @param actionItem {@link SpeedDialActionItem} to be appended to this list
      * @param position   index at which the specified element is to be inserted
+     * @return the instance of the {@link FabWithLabelView} if the add was successful, null otherwise.
      */
-    public void addActionItem(SpeedDialActionItem actionItem, int position) {
-        addActionItem(actionItem, position, true);
+    @Nullable
+    public FabWithLabelView addActionItem(SpeedDialActionItem actionItem, int position) {
+        return addActionItem(actionItem, position, true);
     }
 
     /**
@@ -299,13 +348,17 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
      * @param actionItem {@link SpeedDialActionItem} to be appended to this list
      * @param position   index at which the specified element is to be inserted
      * @param animate    true to animate the insertion, false to insert instantly
+     * @return the instance of the {@link FabWithLabelView} if the add was successful, null otherwise.
      */
-    public void addActionItem(SpeedDialActionItem actionItem, int position, boolean animate) {
+    @Nullable
+    public FabWithLabelView addActionItem(SpeedDialActionItem actionItem, int position, boolean animate) {
         FabWithLabelView oldView = findFabWithLabelViewById(actionItem.getId());
         if (oldView != null) {
-            replaceActionItem(oldView.getSpeedDialActionItem(), actionItem);
+            return replaceActionItem(oldView.getSpeedDialActionItem(), actionItem);
         } else {
-            FabWithLabelView newView = createNewFabWithLabelView(actionItem);
+            FabWithLabelView newView = actionItem.createFabWithLabelView(getContext());
+            newView.setOrientation(getOrientation() == VERTICAL ? HORIZONTAL : VERTICAL);
+            newView.setOnActionSelectedListener(mOnActionSelectedProxyListener);
             int layoutPosition = getLayoutPosition(position);
             addView(newView, layoutPosition);
             mFabWithLabelViews.add(position, newView);
@@ -316,6 +369,7 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
             } else {
                 newView.setVisibility(GONE);
             }
+            return newView;
         }
     }
 
@@ -365,11 +419,11 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
      *
      * @param newActionItem {@link SpeedDialActionItem} to use for the replacement
      * @param position      the index of the {@link SpeedDialActionItem} to be replaced
-     * @return true if this list contained the specified element
+     * @return the instance of the new {@link FabWithLabelView} if the replace was successful, null otherwise.
      */
-    public boolean replaceActionItem(SpeedDialActionItem newActionItem, int position) {
-        return replaceActionItem(mFabWithLabelViews.get(position).getSpeedDialActionItem(),
-                newActionItem);
+    @Nullable
+    public FabWithLabelView replaceActionItem(SpeedDialActionItem newActionItem, int position) {
+        return replaceActionItem(mFabWithLabelViews.get(position).getSpeedDialActionItem(), newActionItem);
     }
 
     /**
@@ -377,25 +431,25 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
      *
      * @param oldSpeedDialActionItem the old {@link SpeedDialActionItem} to remove
      * @param newSpeedDialActionItem the new {@link SpeedDialActionItem} to add
-     * @return true if this list contained the specified element
+     * @return the instance of the new {@link FabWithLabelView} if the replace was successful, null otherwise.
      */
-    public boolean replaceActionItem(@Nullable SpeedDialActionItem oldSpeedDialActionItem,
-                                     SpeedDialActionItem newSpeedDialActionItem) {
+    @Nullable
+    public FabWithLabelView replaceActionItem(@Nullable SpeedDialActionItem oldSpeedDialActionItem,
+                                              SpeedDialActionItem newSpeedDialActionItem) {
         if (oldSpeedDialActionItem == null) {
-            return false;
+            return null;
         } else {
             FabWithLabelView oldView = findFabWithLabelViewById(oldSpeedDialActionItem.getId());
             if (oldView != null) {
                 int index = mFabWithLabelViews.indexOf(oldView);
                 if (index < 0) {
-                    return false;
+                    return null;
                 }
                 removeActionItem(findFabWithLabelViewById(newSpeedDialActionItem.getId()), null, false);
                 removeActionItem(findFabWithLabelViewById(oldSpeedDialActionItem.getId()), null, false);
-                addActionItem(newSpeedDialActionItem, index, false);
-                return true;
+                return addActionItem(newSpeedDialActionItem, index, false);
             } else {
-                return false;
+                return null;
             }
         }
     }
@@ -677,33 +731,25 @@ public class SpeedDialView extends LinearLayout implements CoordinatorLayout.Att
         floatingActionButton.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(final View view) {
-                if (!isOpen() && !mFabWithLabelViews.isEmpty()) {
-                    open();
-                } else {
+                if (isOpen()) {
                     if (mOnChangeListener == null || !mOnChangeListener.onMainActionSelected()) {
                         close();
                     }
+                } else {
+                    open();
                 }
             }
         });
         return floatingActionButton;
     }
 
-    private FabWithLabelView createNewFabWithLabelView(SpeedDialActionItem speedDialActionItem) {
-        FabWithLabelView newView;
-        int theme = speedDialActionItem.getTheme();
-        if (theme == RESOURCE_NOT_SET) {
-            newView = new FabWithLabelView(getContext());
-        } else {
-            newView = new FabWithLabelView(new ContextThemeWrapper(getContext(), theme), null, theme);
-        }
-        newView.setSpeedDialActionItem(speedDialActionItem);
-        newView.setOrientation(getOrientation() == VERTICAL ? HORIZONTAL : VERTICAL);
-        newView.setOnActionSelectedListener(mOnActionSelectedProxyListener);
-        return newView;
-    }
-
     private void toggle(boolean show, boolean animate) {
+        if (show && mFabWithLabelViews.isEmpty()) {
+            show = false;
+            if (mOnChangeListener != null) {
+                mOnChangeListener.onMainActionSelected();
+            }
+        }
         if (isOpen() == show) {
             return;
         }